vql/linux/bpflib: add helper package for bpf plugins
We're about to add several plugins that operate using BPF.  This package
adds some helpers that will be shared among them.

We use the libbpfgo module to implement the BPF interface.  The libbpfgo
project wants to build its own libbpf and statically link with it.  In order
to build the embedded libbpf, we need to incorporate libbpfgo as a submodule.

There are external requirements to build the BPF programs:
- bpftool (unless the builder provides a pregenerated vmlinux.h)
- clang >= 10
- llvm-strip

Since these external requirements are new, the BPF plugins will only be
enabled if the above requirements are met.  If they are unmet, a message will be
printed informing the user about the missing tools and the build will
continue.  The plugins can be made mandatory by setting the environment
variable: BUILD_BPF_PLUGINS=1.