Implement TLS certificate pinning and Fallback Address (#2585)
This PR adds two features to Velociraptor:
* TLS Certificate Pinning: if one specifies the following snippet in the
Client configuration, Velocirpator enables certificate pinning for both
it's API connections and HTTP(S) connections established by
vql.networking.GetHTTPTransport.
```
Client:
  certificate_hashes:
    - 10b5e383aec607420511a30ae520d500e81a50986e3f83a8fb407746060a916c
    - c1033930d035804c8cc74be7e797068fc00973344c9637869f27054088e75a87
  # only if the following value is set to true, strict cert pinning is enforced and 
  # connections are aborted if no fingerprints match; otherwise the fingerprints 
  # are only used to implement some kind of 'early-accept' for connections to 
  # bypass certificate validation for self-signed certificates
  force_certificate_pinning: true
```
* Fallback Addresses: we experienced several situations where we could
not use Velicoraptor in customer's networks because of DNS being
unavailable. This PR introduces the concept of fallback addresses which
Velociraptor can fall back to when the connection to the primary address
could not be established. The idea is roughly as follows: first, try to
connecto to the original target; if that did not work, try the fallback
address (if configured). Considering the example below, whenever a
connection to `my-velociraptor-server.com:443` could not be established,
`123.123.123.123:443` would be tried as a fallback.
```
Client:
  fallback_addresses:
    "my-velociraptor-server.com:443": "123.123.123.123:443"
```

Please note: both the certificate pinning and the fallback addresses
currently only apply to connections which are established by
`vql.networking.GetHttpTransport`. Other connections don't take the
fallback addresses into account. I did intentionally not overwrite
`net.http.DefaultTransport` as I am not able to tell whether or not this
would break other parts of the code.

During the implementation of the TLS cert pinning, I experienced that
the API connections used a different TLS config and HTTP Transport than
the VQL. I took the chance to align both code paths so that the
connection settings don't differ between components of Velociraptor.
This felt like a good idea at that time but may turn out to not be
beneficial. Please let me know whether or not I should revert this and
leave the API comms untouched.

---------

Co-authored-by: Mike Cohen <mike@velocidex.com>