logscale: rework contexts
We don't need to use errgroups for the workers since none of them return
an error.  This commit switches them back to regular cancelable
contexts rooted in the plugin context.  When the plugin is terminated or
the queue is closed, we'll flush the listener and allow the events to be
posted (or fail fast) before returning.